% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints.R
\name{checkConstraints}
\alias{checkConstraints}
\title{Check for spatial constraints}
\usage{
checkConstraints(m, row, col, mode)
}
\arguments{
\item{m}{matrix}

\item{row}{current selected row in the matrix m}

\item{col}{current selected column in the matrix m}

\item{mode}{spatial constraint}
}
\value{
A vector containing the neighbors of element (row,col) of the
matrix m.
}
\description{
Finds the neighbors of the current element (row, col) in the
matrix m, depending on the chosen constraint pattern. Currently, there are
only 3 valid patterns (NS, WE and NEWS)
}
