% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wrapper_wpm.R
\name{checkWpmInputs}
\alias{checkWpmInputs}
\title{Check the inputs for the wrapper function}
\usage{
checkWpmInputs(
  user_df,
  plate_dims,
  nb_plates,
  spatial_constraint,
  max_iteration
)
}
\arguments{
\item{user_df}{expected dataframe, returns adapted message error}

\item{plate_dims}{expected list of plate dimensions (rows and columns)}

\item{nb_plates}{expected number of plates}

\item{spatial_constraint}{expected character for spatial constraint}

\item{max_iteration}{expected number of iterations}
}
\value{
returns an error message if a problem is found with some parameter.
}
\description{
Checks if all the inputs given to the function WrapperWPM are
correct and intercompatible.
}
