% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_imports.R
\name{convertCSV}
\alias{convertCSV}
\title{Convert a CSV File into a dataframe}
\usage{
convertCSV(dt_path, head = FALSE, qt = "", sep = "", row_names = FALSE)
}
\arguments{
\item{dt_path}{file path.}

\item{head}{logical value, indicates if the file has a header on it's first
line.}

\item{qt}{character, the set of quoting characters. To disable quoting
altogether, use quote = "".}

\item{sep}{character, the field separator character. Values on each line
of the file are separated by this character.}

\item{row_names}{logical value, does the file have rownames?}
}
\value{
a dataframe containing 3 fields: Sample, Group and ID.
}
\description{
This function converts a CSV into a dataframe to make it
usable by the shiny application of wpm as well as by the wrapper function
(version of wpm in command line)
}
\examples{
test <- data.frame("Sample" = c("s1","s2","s3","s4"),
                   "Group" = c("A","A","B", "C"))
tf <- tempfile()
write.csv2(test, tf, row.names = FALSE)
convertCSV(tf, head = TRUE, sep = ";")

# if there are row names in the CSV file
write.csv2(test, tf)
convertCSV(tf, head = TRUE, sep = ";", row_names = TRUE)
}
