% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_conversion.R
\name{convertVector2Df}
\alias{convertVector2Df}
\title{Convert a vector of plate coordinates into a dataframe}
\usage{
convertVector2Df(chr_wells, max_Row, max_Col, status = NA)
}
\arguments{
\item{chr_wells}{character string containing the wells}

\item{max_Row}{integer, maximal number of lines in the plate}

\item{max_Col}{integer, maximal number of columns in the plate}

\item{status}{character, the status of the wells}
}
\value{
result, dataframe containing wells coordinates
}
\description{
Function converting the format of "Letter-Digit" coordinates
into a dataframe containing these coordinates in Row, Column.
}
\examples{
# convert the vector of well coordinates into a dataframe
convertVector2Df("A1,C2,A3,B12,C42",3,42,"specify_status")

# supports uppercase / lowercase letters
convertVector2Df("a1,C2,A3,b12,C42",3,42,"specify_status")

}
