% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{Fictitious clinical data for demonstration.}
\format{
A data frame with 193 rows and 7 variables

\describe{
    \item{samples}{the samples to be analyzed representing fictitious patients.}
    \item{age}{age of the patients under 5 age groups, in years (20-30;30-40;40-50;50-60;60-70)}
    \item{gender}{gender of the patients, F (for Female) and M (for Male)}
    \item{treatment}{the treatment each patient received, Ctrl (Control), treatment A, treatment B and treatment C}
    \item{diabetes}{presence of diabetes, 0 (no) and 1 (yes) }
    \item{gender-treatment}{A combination between the gender and treatment fields}
    \item{age-diabetes}{A combination between the age and the diabetes fields}
}
}
\usage{
data(data_test)
}
\description{
A demo dataset containing the age and other attributes of 193
fictitious patients. It aims to help the user to test the shiny application
of the wpm package.
}
\author{
Helene Borges \email{borges.helene.sophie@gmail.com}
}
\keyword{datasets}
