% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_blank_coordinates.R
\name{defineBlankCoords}
\alias{defineBlankCoords}
\title{Determines blanks wells coordinates on a plate}
\usage{
defineBlankCoords(p_lines, p_cols, mod = "none", start_blank)
}
\arguments{
\item{p_lines}{number of rows on the plate}

\item{p_cols}{number of columns on the plate}

\item{mod}{character, can be "none", "by_column", "by_row" or "checkerboard"}

\item{start_blank}{character, "even" means that the blanks will be
positioned on the even rows of the plate. Otherwise, they will be
positioned on the odd rows.}
}
\value{
a dataframe containing the blank wells with their coordinates.
}
\description{
function to place the blanks on the plate according to the
selected mode: it generates a dataframe containing the row and column
coordinates for each blank.
}
