% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_buffer_coordinates.R
\name{defineBufferCoords}
\alias{defineBufferCoords}
\title{Determines buffer wells coordinates on a plate}
\usage{
defineBufferCoords(p_lines, p_cols, mod = "none", start_buffer)
}
\arguments{
\item{p_lines}{number of rows on the plate}

\item{p_cols}{number of columns on the plate}

\item{mod}{character, can be "none", "by_column", "by_row" or "checkerboard"}

\item{start_buffer}{character, "even" means that the buffers will be
positioned on the even rows of the plate. Otherwise, they will be
positioned on the odd rows.}
}
\value{
a dataframe containing the buffer wells with their coordinates.
}
\description{
function to place the buffer solutions on the plate according 
to the selected mode: it generates a dataframe containing the row and column
coordinates for each buffer solution.
}
