% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints.R
\name{findNEWSneighbors}
\alias{findNEWSneighbors}
\title{Find the 4 cardinal neighbors of an element of a matrix}
\usage{
findNEWSneighbors(m, i, j)
}
\arguments{
\item{m, }{matrix}

\item{i, }{integer, line index in the matrix}

\item{j, }{integer, column index in the matrix}
}
\value{
A vector containing the North, East, West and South neighbors of
the element (i,j) of the matrix being processed.
}
\description{
Function for spatial contraints: the North, East, West and
South neighbors of the current element (i,j) of the matrix m.
}
