% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints.R
\name{findNSneighbors}
\alias{findNSneighbors}
\title{Find the top and bottom neighbors of an element of a matrix}
\usage{
findNSneighbors(m, i, j)
}
\arguments{
\item{m, }{matrix}

\item{i, }{integer, line index in the matrix}

\item{j, }{integer, column index in the matrix}
}
\value{
A vector containing the North and South neighbors of the element
(i,j) of the matrix being processed.
}
\description{
Function for spatial constraint that only looks for North and
South
}
