% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints.R
\name{findWEneighbors}
\alias{findWEneighbors}
\title{Find the left and right neighbors of an element of a matrix}
\usage{
findWEneighbors(m, i, j)
}
\arguments{
\item{m, }{matrix}

\item{i, }{integer, line index in the matrix}

\item{j, }{integer, column index in the matrix}
}
\value{
A vector containing the West and East neighbors of the element
(i,j) of the matrix being processed
}
\description{
Function for spatial constraint that only looks for West and
East neighbors of the current element (i,j) of the matrix.
}
