% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_import.R
\name{mod_data_import_server}
\alias{mod_data_import_server}
\title{data_import Server Function}
\usage{
mod_data_import_server(input, output, session, stringsAsFactors)
}
\arguments{
\item{input, output, session}{Internal shiny parameters}

\item{stringsAsFactors}{Boolean, argument passed to the function `read.csv2`}
}
\value{
dataframe containing the data to place on the plates plan(s).
}
\description{
server part of the data import module. Allows to browse a file
in CSV, text or TSV format, create a dataframe and set the column names.
}
