% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plate_specifications.R
\name{mod_plate_specifications_ui}
\alias{mod_plate_specifications_ui}
\title{plate_specifications UI Function}
\usage{
mod_plate_specifications_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
a Build Page layout
}
\description{
A shiny Module allowing to process the parameters specified
by the user in order to generate the plate plans afterwards.
}
