% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_random_walk.R
\name{randomWalk}
\alias{randomWalk}
\title{Random walk of the matrix to fill}
\usage{
randomWalk(m, toVisit, d, constraint)
}
\arguments{
\item{m}{is a matrix corresponding to the plate to be filled.}

\item{toVisit}{contains the wells in form "A1", and contains only the 
wells authorized to be filled in}

\item{d}{is the dataframe containing the data supplied by the user.}

\item{constraint}{character string corresponding to the spatial constraint 
selected by the user}
}
\value{
a dataframe corresponding to the user-supplied data. This dataframe 
is an updated version, where the columns `Row` and `Column` are filled with 
the coordinates of the chosen well. If no solution is found for the current
selected well, then this function returns 1.
}
\description{
Returns the user dataframe updated after choosing randomly a well on the 
plate (matrix) and randomly choosing a sample ID that satisfies all the 
constraints.
}
