% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resample.R
\name{resample}
\alias{resample}
\title{Randomly take a number of elements in a vector}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{is a vector}

\item{...}{parameters given to the function sample.int}
}
\value{
a vector of length equal to size parameter.
}
\description{
This function allows to pick up the last element in a vector 
when the parameter size is equal to 1. Passes parameters to `sample.int` 
like size.
}
