% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm.R
\name{wpm}
\alias{wpm}
\title{Run the Shiny Application of Well Plate Maker}
\usage{
wpm(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\value{
a shiny application object with golem options
}
\description{
Run the Shiny Application of Well Plate Maker
}
\examples{
if(interactive()) {wpm()}
}
