% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_join_dataframes.R
\name{joinDataframes}
\alias{joinDataframes}
\title{Binds multiple dataframes together}
\usage{
joinDataframes(
  forbidden_w = NULL,
  buffer_w = NULL,
  fixed_w = NULL,
  nb_samples,
  totalNbWells,
  nb_p
)
}
\arguments{
\item{forbidden_w}{dataframe, the forbidden wells}

\item{buffer_w}{datarame, the buffer wells}

\item{fixed_w}{dataframe, the quality control wells}

\item{nb_samples}{numeric, the number of samples to place using the
backtracking algorithm.}

\item{totalNbWells, }{numeric, the total number of wells that can be filled.}

\item{nb_p}{numeric, number of plates to fill}
}
\value{
a dataframe containing all the special wells
}
\description{
Function that merges dataframes that contain wells of
different types. To do this, it verifies that all the conditions provided
are compatible with each other in order to be able to launch WPM on this
data.
}
