% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm.R
\name{wpm}
\alias{wpm}
\title{wpm}
\usage{
wpm()
}
\value{
launches the WPM app in a new window in the default browser
}
\description{
Well-Plate Maker : a GUI for building well-plates plans.
}
\examples{
wpm()
}
\author{
Helene  Borges
}
